import/requirements.txt
cassandra-driver
pandas
redis
